requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langchain_ollama
langserve
fastapi
uvicorn
sse_starlette